backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.30
alembic>=1.13.1

# PDF Parsing
pdfplumber>=0.11.0

# Data Processing
pandas>=2.2.3
openpyxl>=3.1.2
xlrd>=2.0.1

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Utilities
python-dateutil>=2.9.0
